+++ src/browser-lib/fixtures/subjectListing.ts
@@ -1,5 +1,5 @@
 import { SubjectListingData } from "../../node-lib/curriculum-api";
 import subjectListingFixture from "../../node-lib/curriculum-api/fixtures/subjectListing.fixture";
 
-export const subjects: SubjectListingData["programmesAvailable"] =
-  subjectListingFixture().programmesAvailable;
+export const subjects: SubjectListingData["subjects"] =
+  subjectListingFixture().subjects;