+++ src/components/SubjectCardList/SubjectCardList.tsx
@@ -1,29 +1,35 @@
 import { FC } from "react";
 
-import { ProgrammesBySubject } from "../../pages/beta/[viewType]/key-stages/[keyStageSlug]/subjects";
+import { SubjectListingPageProps } from "../../pages/beta/[viewType]/key-stages/[keyStageSlug]/subjects";
 import Grid, { GridArea } from "../Grid";
 
 import SubjectCardListItem from "./SubjectCardListItem";
 
-export type SubjectCardListProps = {
-  subjects: ProgrammesBySubject[];
+export type SubjectCardListProps = Omit<
+  SubjectListingPageProps,
+  "subjectsUnavailable"
+> & {
   isAvailable: boolean;
 };
 
 const SubjectCardList: FC<SubjectCardListProps> = ({
   subjects,
   isAvailable,
+  keyStageSlug,
+  keyStageTitle,
 }) => {
   return (
     <Grid $rg={16} $cg={16} $gridAutoRows={"1fr"} $mb={72}>
-      {subjects.map((programmes) => {
+      {subjects.map((subject) => {
         return (
           <GridArea
-            key={`subject-list-item-${programmes[0].subjectSlug}`}
+            key={`subject-list-item-${subject[0].subjectSlug}`}
             $colSpan={[6, 3, 2]}
           >
             <SubjectCardListItem
-              programmes={programmes}
+              subject={subject}
+              keyStageSlug={keyStageSlug}
+              keyStageTitle={keyStageTitle}
               isAvailable={isAvailable}
             />
           </GridArea>